Extract NavigationEntityItems and add getKey
