+++ packages/edit-site/src/components/left-sidebar/navigation-panel/page-switcher/index.js
@@ -1,52 +1,13 @@
 /**
  * WordPress dependencies
  */
-import { Fragment } from '@wordpress/element';
 import { getPathAndQueryString } from '@wordpress/url';
-import { useSelect } from '@wordpress/data';
-import {
-	__experimentalNavigationGroup as NavigationGroup,
-	__experimentalNavigationItem as NavigationItem,
-} from '@wordpress/components';
+import { __experimentalNavigationGroup as NavigationGroup } from '@wordpress/components';
 
-const NavigationEntityItems = ( props ) => {
-	const {
-		kind,
-		name,
-		query = {},
-
-		renderItem,
-	} = props;
-
-	const entities = useSelect(
-		( select ) => {
-			const { getEntityRecords } = select( 'core' );
-			return getEntityRecords( kind, name, query );
-		},
-		[ kind, name, query ]
-	);
-
-	if ( ! entities ) {
-		return null;
-	}
-
-	return entities.map( ( entity ) => {
-		const key = `${ kind }-${ name }-${ entity.id }`;
-
-		return (
-			<Fragment key={ key }>
-				{ renderItem( {
-					ItemComponent: NavigationItem,
-					entity,
-					item: key,
-					props: {
-						item: key,
-					},
-				} ) }
-			</Fragment>
-		);
-	} );
-};
+/**
+ * Internal dependencies
+ */
+import NavigationEntityItems from './navigation-entity-items';
 
 export default function PageSwitcher( { onChangePage } ) {
 	const onPageSelect = ( { type, slug, link, id }, item ) => {
@@ -64,16 +25,19 @@ export default function PageSwitcher( { onChangePage } ) {
 		);
 	};
 
+	const keyExtractor = ( { entity: { link } } ) =>
+		getPathAndQueryString( link );
+
 	return (
 		<>
 			<NavigationGroup title="Pages">
 				<NavigationEntityItems
 					kind="postType"
 					name="page"
+					getKey={ keyExtractor }
 					renderItem={ ( { ItemComponent, entity, item, props } ) => (
 						<ItemComponent
 							{ ...props }
-							item={ getPathAndQueryString( entity.link ) }
 							title={ entity.title.rendered }
 							onClick={ () => onPageSelect( entity, item ) }
 						/>
@@ -85,10 +49,10 @@ export default function PageSwitcher( { onChangePage } ) {
 				<NavigationEntityItems
 					kind="taxonomy"
 					name="category"
+					getKey={ keyExtractor }
 					renderItem={ ( { ItemComponent, entity, item, props } ) => (
 						<ItemComponent
 							{ ...props }
-							item={ getPathAndQueryString( entity.link ) }
 							title={ entity.name }
 							onClick={ () => onPageSelect( entity, item ) }
 						/>
@@ -100,10 +64,10 @@ export default function PageSwitcher( { onChangePage } ) {
 				<NavigationEntityItems
 					kind="postType"
 					name="post"
+					getKey={ keyExtractor }
 					renderItem={ ( { ItemComponent, entity, item, props } ) => (
 						<ItemComponent
 							{ ...props }
-							item={ getPathAndQueryString( entity.link ) }
 							title={ entity.title.rendered }
 							onClick={ () => onPageSelect( entity, item ) }
 						/>
